`asoc-delete` need not return original list
Previously, `asoc-delete` guaranteed that it would return the same list you
passed in, unless the result was nil.

This restriction makes the code less efficient and more verbose: the function
cannot simply return the tail when the head is to be removed, it must reassign
car and cdr.

However, there is little use for this guarantee. The case of a nil result must
still be handled, so it's best to use the function by assigning the original
alist to the result of the `asoc-delete` call - in which case the guarantee is
unnecessary.

Note: the guarantee was not even covered by the unit tests, so no changes are
needed to the test file.

* Remove the guarantee from the docstring, and replace with a statement that
the function may destructively modify the alist.
* Rewrite the implementation more concisely and efficiently.