Remove `keep-last` parameter from `asoc-uniq`, update docs & tests
The option to keep the last occurrence of a key is unlikely to be needed.
If it is, the alist can simply be reversed. Removing the option simplifies
`asoc-uniq`, making it equivalent to `asoc-merge` with a single argument.
It also leaves the option for expanding the interface later.

Also, update `asoc-merge` docstring and move `asoc-uniq` before `asoc-merge`.

Increase version to 0.4.0.