Remove asoc-delete! and replace with asoc-pop! and asoc-dissoc
`asoc-delete!` doesn't have the nicest interface. It mutates the alist (like
the `delete` family of functions for lists), so the original alist is not
preserved. However, it cannot be guaranteed to be left as the result alist
either (since it's impossible to mutate a list into nil). This is done for
efficiency, but the extra efficiency is probably not needed (and has not even
been measured). Thus, remove this function for now. It can be restored if a
highly efficient deletion function is ever required.

Replace with `asoc-pop!` for the case where only the first matching association
is removed. This sets the alist to the new value and returns the removed
association. In the remove-all case, replace with the pure function
`asoc-dissoc`, which accepts multiple key arguments and returns a result with
all occurrences of them removed.